workflows/requirements.txt
redis==5.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
langchain==0.1.9
langchain-community==0.0.38
langchain-groq==0.1.3
langgraph==0.0.26
pgvector==0.2.5
sentence-transformers==2.7.0
